增加实时数据接口

+++ src/main/java/cn/relaxtech/doomsday/bi/boot/entity/bi/BiGeneralData.java
@@ -22,7 +22,7 @@ public class BiGeneralData {
     private Double WRR;
     private Double HMRR;
     private Double MRR;
-    private Integer PAC;
+    private Double PAC;
     private Integer PUC;
     private Double ARPPU;
     private Double ARPU;
